float->decimal частично

+++ db/db_init.sql
@@ -50,9 +50,9 @@ CREATE TABLE IF NOT EXISTS ions_library (
     id INT AUTO_INCREMENT PRIMARY KEY,
     lattice_type_id INT NOT NULL,
     substance_id INT NOT NULL,
-    atom_site_fract_x FLOAT,
-    atom_site_fract_y FLOAT,
-    atom_site_fract_z FLOAT,
+    atom_site_fract_x DECIMAL(10, 6),
+    atom_site_fract_y DECIMAL(10, 6),
+    atom_site_fract_z DECIMAL(10, 6),
     FOREIGN KEY (lattice_type_id) REFERENCES lattice_type(id) ON DELETE CASCADE,
     FOREIGN KEY (substance_id) REFERENCES substances(id) ON DELETE CASCADE
 );
